Create Week-4-4.3
A friend function is a non-member function that has access to the private and protected members of 
a class. 
It is declared inside the class using the friend keyword but defined outside the class. 
Friend functions are useful when two or more classes need to share data or when external functions 
need controlled access to private members.