Create Week-5-5.3
In C++, a copy constructor is a special constructor used to create a new object as a copy of an 
existing object. 
It takes a reference to another object of the same class as its parameter. 
The copy constructor is invoked: 
• When an object is initialized from another object (Rectangle rect2 = rect1;). 
• When an object is passed by value to a function. 
• When an object is returned by value from a function. 
By defining a custom copy constructor, we can control how the copying of object data happens, 
instead of relying on the compiler-generated default